% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_by_group.R
\name{mice_by_group}
\alias{mice_by_group}
\title{Multivariate Imputation by Chained Equations within Groups}
\usage{
mice_by_group(
  Data,
  groupvar,
  miceArgs = list(m = 10, maxit = 10, seed = 123, printFlag = FALSE)
)
}
\arguments{
\item{Data}{A data frame containing all variables (columns) to be used in
the MICE equations and a column with the grouping variable.}

\item{groupvar}{The column name in \code{Data} for the grouping variable. If
this column is not of the class 'factor', it will be converted to a factor
with the default order of factor levels.}

\item{miceArgs}{These are arguments to be passed on to the \code{mice}
function from the 'mice' package. These include the number of imputed data
frames to be generated \code{m} (default = 10), the maximum number of MICE
iterations \code{maxit} (default = 10), a random seed to be set for the MICE
for reproducibility \code{seed} (default = 123), and a Boolean option for
verbose printing \code{printFlag} (default = FALSE). All other \code{mice}
function arguments default to the same values as in the 'mice' package (see
\code{?mice::mice} for details). Note that the default \code{m} and
\code{maxit} for \code{mice} are both 5, though I set them to 10 for this
function for more iterations and better convergence, respectively.}
}
\value{
\code{mice_by_group} returns a list of imputed data frames.
}
\description{
\code{mice_by_group} takes a data frame with a grouping column and columns
to be used in the MICE imputation and returns a list of data frames with
multiply imputed data.
}
\examples{
library(mice)

nhanes$Group <- as.factor(rep(LETTERS[1:2], times = c(12, 13)))
implist <-  mice_by_group(nhanes, "Group", list(m=5,maxit=5,seed=2))

}
